requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
altair
plotly
biopython
requests
